+++ HISTORY.rst
@@ -10,6 +10,7 @@ dev
 * Provide a matching csvpatch command applying diffs.
 * Add a man page and docs for csvpatch.
 * Use exit codes to indicate difference.
+* Add a --quiet option to csvdiff.
 
 0.2.0 (2014-12-30)
 ~~~~~~~~~~~~~~~~~~